refactor(bulk-editor): align the table component with the stack conventions
Sources the field set ids from the central constants, removes the
PropTypes per the React 19 convention, and sizes the loading skeleton
to the new shared PAGE_SIZE so the loading state matches a full page.
