+++ packages/js/src/bulk-editor/constants.js
@@ -5,6 +5,9 @@ export const STORE_NAME = "@yoast/bulk-editor";
 
 export const ROOT_ID = "yoast-seo-bulk-editor";
 
+// How many rows the bulk editor shows per page; also sizes the loading skeleton.
+export const PAGE_SIZE = 20;
+
 // The field sets behind the two tabs: which editable fields the table shows.
 export const FIELD_SET_SEARCH = "search";
 export const FIELD_SET_SOCIAL = "social";
